[#61] refactor: orderService common 모듈로 이동
