+++ kyumall-client/src/test/java/com/kyumall/kyumallclient/order/OrdersIntegrationTest.java
@@ -7,6 +7,8 @@ import com.kyumall.kyumallclient.AuthTestUtil;
 import com.kyumall.kyumallclient.IntegrationTest;
 import com.kyumall.kyumallclient.product.ProductIntegrationTest;
 import com.kyumall.kyumallcommon.factory.MemberFactory;
+import com.kyumall.kyumallcommon.order.dto.CreateOrderRequest;
+import com.kyumall.kyumallcommon.order.dto.ProductIdAndCount;
 import com.kyumall.kyumallcommon.pay.PayOpenFeign;
 import com.kyumall.kyumallcommon.pay.PayResponse;
 import com.kyumall.kyumallcommon.factory.ProductFactory;
@@ -15,7 +17,7 @@ import com.kyumall.kyumallcommon.fixture.product.ProductFixture;
 import com.kyumall.kyumallcommon.member.entity.Member;
 import com.kyumall.kyumallcommon.order.entity.Orders;
 import com.kyumall.kyumallcommon.order.repository.OrderRepository;
-import com.kyumall.kyumallcommon.order.vo.OrderStatus;
+import com.kyumall.kyumallcommon.order.entity.OrderStatus;
 import com.kyumall.kyumallcommon.product.product.Product;
 import com.kyumall.kyumallcommon.product.stock.StockRepository;
 import io.restassured.RestAssured;
